Update result display to support better error messages
