--- src/main/java/seedu/address/ui/ResultDisplay.java
@@ -1,28 +1,0 @@
-package seedu.address.ui;
-
-import static java.util.Objects.requireNonNull;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.TextArea;
-import javafx.scene.layout.Region;
-
-/**
- * A ui for the status bar that is displayed at the header of the application.
- */
-public class ResultDisplay extends UiPart<Region> {
-
-    private static final String FXML = "ResultDisplay.fxml";
-
-    @FXML
-    private TextArea resultDisplay;
-
-    public ResultDisplay() {
-        super(FXML);
-    }
-
-    public void setFeedbackToUser(String feedbackToUser) {
-        requireNonNull(feedbackToUser);
-        resultDisplay.setText(feedbackToUser);
-    }
-
-}
