+++ src/main/resources/view/DarkTheme.css
@@ -337,6 +337,16 @@
     -fx-background-radius: 0;
 }
 
+#inputFormat .content {
+    -fx-background-color: transparent, #383838, transparent, #383838;
+    -fx-background-radius: 0;
+}
+
+#inputExample .content {
+    -fx-background-color: transparent, #383838, transparent, #383838;
+    -fx-background-radius: 0;
+}
+
 #tags {
     -fx-hgap: 7;
     -fx-vgap: 3;
