+++ src/main/resources/view/ResultPanel.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.VBox?>
+
+<StackPane fx:id="placeHolder" styleClass="pane-with-border" xmlns="http://javafx.com/javafx/8"
+           xmlns:fx="http://javafx.com/fxml/1">
+    <VBox>
+      <TextArea fx:id="resultDisplay" editable="false" styleClass="result-display" wrapText="true"/>
+      <TextArea fx:id="inputFormat" editable="false" styleClass="result-display" wrapText="true"/>
+      <TextArea fx:id="inputExample" editable="false" styleClass="result-display" wrapText="true"/>
+    </VBox>
+</StackPane>
